cobalt-171 Add parsing of "this". Moved parsers list to Parser class.
