+++ src/main/scala/compiler/structure/parsers/packages/PackageParser.scala
@@ -30,7 +30,7 @@ class PackageParser extends Parser[PackageBlock] {
     *
     * @return
     */
-  override def getRegexs: List[String] = List("package[ ]+[a-zA-Z][a-zA-Z0-9]*(\\.[a-zA-Z][a-zA-Z0-9]*)*")
+  override def getRegexs: List[String] = List("package[ ]+[a-zA-Z][a-zA-Z0-9\\_]*(\\.[a-zA-Z][a-zA-Z0-9\\_]*)*")
 
   def parse(superBlock: Block, tokenizer: Tokenizer): PackageBlock = {
     tokenizer.nextToken // skip "package"
