+++ src/main/scala/compiler/structure/parsers/this_keyword/ThisKeywordParser.scala
@@ -16,30 +16,25 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package compiler.structure.parsers.operators.assignment.bit
+package compiler.structure.parsers.variable
 
-import compiler.structure.blocks.operators.assignment.bit.LeftShiftAssignOpBlock
-import compiler.structure.parsers.Parsers
+import compiler.structure.blocks.Block
+import compiler.structure.blocks.this_keyword.ThisKeywordBlock
+import compiler.structure.parsers.Parser
 import compiler.tokenizer.Tokenizer
-import org.junit.runner.RunWith
-import org.scalatest.junit.JUnitRunner
-import org.scalatest.{BeforeAndAfter, FunSuite}
 
-@RunWith(classOf[JUnitRunner])
-class LeftShiftOpParserTest() extends FunSuite with BeforeAndAfter {
+class ThisKeywordParser extends Parser[ThisKeywordBlock] {
 
-  val parsers = Parsers.parsers
-
-  val lines = List(
-    "<<="
+  /**
+    * A list of all regular expressions
+    *
+    * @return
+    */
+  override def getRegexs: List[String] = List(
+    "this"
   )
 
-  test("Block creation test") {
-    for (line <- lines) {
-      val parseable = parsers.filter(p => p.shouldParse(line))
-      assert(!parseable.isEmpty)
-      assert(parseable.head.parse(null, new Tokenizer(line)).isInstanceOf[LeftShiftAssignOpBlock])
-    }
+  def parse(superBlock: Block, tokenizer: Tokenizer): ThisKeywordBlock = {
+    new ThisKeywordBlock(superBlock)
   }
-
 }
