+++ src/test/scala/compiler/structure/parsers/constants/CharConstantParserTest.scala
@@ -20,15 +20,15 @@ package compiler.structure.parsers.constants
 
 import compiler.structure.blocks.Block
 import compiler.structure.blocks.constants.CharConstantBlock
+import compiler.structure.parsers.Parsers
 import compiler.tokenizer.Tokenizer
-import compiler.utilities.Constants
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{BeforeAndAfter, FunSuite}
 
 @RunWith(classOf[JUnitRunner])
 class CharConstantParserTest extends FunSuite with BeforeAndAfter {
-  val parsers = Constants.parsers
+  val parsers = Parsers.parsers
 
   val lines = List(
     "'a'",
