+++ src/test/scala/compiler/structure/parsers/constants/ShortConstantParserTest.scala
@@ -20,9 +20,8 @@ package compiler.structure.parsers.constants
 
 import compiler.structure.blocks.Block
 import compiler.structure.blocks.constants.ShortConstantBlock
-import compiler.structure.parsers.Parser
+import compiler.structure.parsers.{Parser, Parsers}
 import compiler.tokenizer.Tokenizer
-import compiler.utilities.Constants
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{BeforeAndAfter, FunSuite}
@@ -30,7 +29,7 @@ import org.scalatest.{BeforeAndAfter, FunSuite}
 @RunWith(classOf[JUnitRunner])
 class ShortConstantParserTest() extends FunSuite with BeforeAndAfter {
 
-  val parsers: List[Parser[_]] = Constants.parsers
+  val parsers: List[Parser[_]] = Parsers.parsers
 
   val lines: List[String] = List(
     "10s",
