+++ src/test/scala/compiler/structure/parsers/operators/assignment/bit/InclOrAssignOpParserTest.scala
@@ -19,8 +19,8 @@
 package compiler.structure.parsers.operators.assignment.bit
 
 import compiler.structure.blocks.operators.assignment.bit.InclOrAssignOpBlock
+import compiler.structure.parsers.Parsers
 import compiler.tokenizer.Tokenizer
-import compiler.utilities.Constants
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{BeforeAndAfter, FunSuite}
@@ -28,7 +28,7 @@ import org.scalatest.{BeforeAndAfter, FunSuite}
 @RunWith(classOf[JUnitRunner])
 class InclOrAssignOpParserTest() extends FunSuite with BeforeAndAfter {
 
-  val parsers = Constants.parsers
+  val parsers = Parsers.parsers
 
   val lines = List(
     "|="
