+++ src/test/scala/compiler/structure/parsers/structures/ObjectDefinitionParserTest.scala
@@ -18,7 +18,7 @@
 
 package compiler.structure.parsers.structures
 
-import compiler.utilities.Constants
+import compiler.structure.parsers.Parsers
 import org.junit.runner.RunWith
 import org.scalatest.junit.JUnitRunner
 import org.scalatest.{BeforeAndAfter, FunSuite}
@@ -26,7 +26,7 @@ import org.scalatest.{BeforeAndAfter, FunSuite}
 @RunWith(classOf[JUnitRunner])
 class ObjectDefinitionParserTest() extends FunSuite with BeforeAndAfter {
 
-  val parsers = Constants.parsers
+  val parsers = Parsers.parsers
 
   val lines = List(
     "new Object()",
